Changed processLine() method: Added delimiter and quotation mark based cell extraction.
Changed accessing of preprocessing to process() method.
Added argument verification for input.

Deleted JavaFX dependencies and classes, switch to javax.swing for tree drawing.
Created new packages for better code distribution.

Overall this commit changes the structure of the code to be more focused on singe line instead of the whole file, which might slow it down but should create a better memory performance.